requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pytest==8.3.3
pytest-mock==3.14.0
python-magic==0.4.27
python-multipart==0.0.17
httpx==0.27.2
ruff==0.8.0
mypy==1.13.0
pre-commit==3.5.0
